# apps/web/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TABSIRA API",
    "description": "Photo to insight, backed by one Quran verse and one hadith.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/insights/{insight_id}": {
      "get": {
        "tags": ["insights"],
        "summary": "One insight, its scripture read from the store",
        "description": "Return the insight: its verse and hadith exactly as stored, the explanation apart.",
        "operationId": "get_insight_insights__insight_id__get",
        "parameters": [
          {
            "name": "insight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,18}$",
              "title": "Insight Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightDetailOut"
                }
              }
            }
          },
          "default": {
            "description": "An error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/insights/{insight_id}/chat": {
      "post": {
        "tags": ["insights"],
        "summary": "Ask one question about the insight (three at most)",
        "description": "Answer a question classified by its content level, grounded in the insight.\n\nThe same `idempotencyKey` returns the same answer and counts once; the\nfourth successful message answers 409 CHAT_LIMIT_REACHED.",
        "operationId": "chat_insights__insight_id__chat_post",
        "parameters": [
          {
            "name": "insight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,18}$",
              "title": "Insight Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatReply"
                }
              }
            }
          },
          "default": {
            "description": "An error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/insights/{insight_id}/action": {
      "post": {
        "tags": ["insights"],
        "summary": "Declare the small step done, or for later",
        "description": "Record «نفّذته» or «سأفعله لاحقًا»: the learner's own statement, never a proof or a reward.",
        "operationId": "declare_action_insights__insight_id__action_post",
        "parameters": [
          {
            "name": "insight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,18}$",
              "title": "Insight Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionOut"
                }
              }
            }
          },
          "default": {
            "description": "An error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/insights/{insight_id}/complete": {
      "post": {
        "tags": ["insights"],
        "summary": "«تمّ»: complete the insight, once",
        "description": "Complete the insight; a second call saves nothing more and answers the same place.",
        "operationId": "complete_insight_insights__insight_id__complete_post",
        "parameters": [
          {
            "name": "insight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,18}$",
              "title": "Insight Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionOut"
                }
              }
            }
          },
          "default": {
            "description": "An error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/insights/{insight_id}/publish": {
      "post": {
        "tags": ["insights"],
        "summary": "Make the insight public, at a link anyone reads",
        "description": "Publish the owner's insight (v2 §18); publishing twice keeps the first time.\n\nA verified account only (decision 25). Refused, 409, for a simulation, for an\ninsight none of whose texts is verified and shown yet, or for a text that\nreads like scripture.",
        "operationId": "publish_insight_insights__insight_id__publish_post",
        "parameters": [
          {
            "name": "insight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,18}$",
              "title": "Insight Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationOut"
                }
              }
            }
          },
          "default": {
            "description": "An error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["insights"],
        "summary": "«إلغاء النشر»: withdraw the insight from the public",
        "description": "Take the insight off its public address, which answers 404 from then on.",
        "operationId": "withdraw_insight_insights__insight_id__publish_delete",
        "parameters": [
          {
            "name": "insight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,18}$",
              "title": "Insight Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationOut"
                }
              }
            }
          },
          "default": {
            "description": "An error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/public/insights/{insight_id}": {
      "get": {
        "tags": ["insights"],
        "summary": "A published insight, as anyone reads it",
        "description": "Return a published insight: the texts shown, exactly as stored, and the platform's words.\n\nNothing of the owner but the public name they chose. Private, withdrawn, missing\nand another account's insights all answer the same 404.",
        "operationId": "get_public_insight_public_insights__insight_id__get",
        "parameters": [
          {
            "name": "insight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,18}$",
              "title": "Insight Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicInsightOut"
                }
              }
            }
          },
          "default": {
            "description": "An error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/tutorial/rain/insights/{slug}": {
      "post": {
        "tags": ["tutorial"],
        "summary": "Keep a copy of a tutorial insight, to complete it and ask about it",
        "description": "Return the caller's copy of the insight (made once), labelled «مثال موثّق مُعدّ».",
        "operationId": "keep_rain_insight_tutorial_rain_insights__slug__post",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9-]{1,63}$",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightDetailOut"
                }
              }
            }
          },
          "default": {
            "description": "An error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ActionIn": {
        "properties": {
          "choice": {
            "$ref": "#/components/schemas/ActionState"
          }
        },
        "type": "object",
        "required": ["choice"],
        "title": "ActionIn",
        "description": "«نفّذته» (`done`) or «سأفعله لاحقًا» (`later`)."
      },
      "ActionOut": {
        "properties": {
          "state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionState"
              },
              {
                "type": "null"
              }
            ]
          },
          "at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "At"
          },
          "means": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Means"
          }
        },
        "type": "object",
        "required": ["state", "at", "means"],
        "title": "ActionOut",
        "description": "What the learner declared about the small step: a statement, never a proof or a reward."
      },
...
      "BBox": {
        "properties": {
          "x": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "X"
          },
          "y": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Y"
          },
          "width": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Width"
          },
          "height": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Height"
          }
        },
        "type": "object",
        "required": ["x", "y", "width", "height"],
        "title": "BBox",
        "description": "A box as ratios of the image; `x` and `y` are its top-left corner."
      },
...
      "ChatIn": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Message"
          },
          "idempotencyKey": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Idempotencykey"
          }
        },
        "type": "object",
        "required": ["message", "idempotencyKey"],
        "title": "ChatIn"
      },
...
      "ChatReply": {
        "properties": {
          "message": {
            "$ref": "#/components/schemas/ChatMessageOut"
          },
          "used": {
            "type": "integer",
            "title": "Used"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "remaining": {
            "type": "integer",
            "title": "Remaining"
          },
          "disclosure": {
            "type": "string",
            "title": "Disclosure"
          }
        },
        "type": "object",
        "required": ["message", "used", "limit", "remaining", "disclosure"],
        "title": "ChatReply"
      },
...
      "CompletionOut": {
        "properties": {
          "insight_id": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$",
            "title": "Insight Id"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Completed At"
          },
          "first_time": {
            "type": "boolean",
            "title": "First Time",
            "description": "False when «تمّ» was already recorded: nothing was saved twice"
          },
          "place": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaceOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "treasure_prepared": {
            "type": "boolean",
            "title": "Treasure Prepared",
            "description": "A hidden treasure waits; it shows on return"
          },
          "badges_earned": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Badges Earned"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/AfterOption"
            },
            "type": "array",
            "title": "Options"
          },
          "suggest_account": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggest Account",
            "description": "Set for a guest after the first completion"
          },
          "disclosure": {
            "type": "string",
            "title": "Disclosure"
          }
        },
        "type": "object",
        "required": [
          "insight_id",
          "completed_at",
          "first_time",
          "place",
          "treasure_prepared",
          "badges_earned",
          "options",
          "suggest_account",
          "disclosure"
        ],
        "title": "CompletionOut"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          },
          "terms_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms Version"
          },
          "privacy_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Version"
          }
        },
        "type": "object",
        "required": ["error", "detail"],
        "title": "ErrorResponse",
        "description": "The body of every error response."
      },
...
      "InsightDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$",
            "title": "Id"
          },
          "scan_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[1-9][0-9]{0,18}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scan Id"
          },
          "origin": {
            "$ref": "#/components/schemas/InsightOrigin"
          },
          "engine": {
            "type": "string",
            "title": "Engine",
            "description": "`pipeline`, `demo` (a declared simulation) or `prepared`"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label",
            "description": "«مثال موثّق مُعدّ» or the simulation notice, when one applies"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "glimpse": {
            "type": "string",
            "title": "Glimpse"
          },
          "anchor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BBox"
              },
              {
                "type": "null"
              }
            ]
          },
          "relation": {
            "$ref": "#/components/schemas/RelationType"
          },
          "relation_label": {
            "type": "string",
            "title": "Relation Label"
          },
          "quran": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InsightQuran"
              },
              {
                "type": "null"
              }
            ]
          },
          "hadith": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InsightHadith"
              },
              {
                "type": "null"
              }
            ]
          },
          "hadith_status": {
            "type": "string",
            "enum": ["shown", "awaiting_verification", "none"],
            "title": "Hadith Status"
          },
          "notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notice",
            "description": "Set when the hadith waits for its dorar.net ruling"
          },
          "pair_complete": {
            "type": "boolean",
            "title": "Pair Complete"
          },
          "explanation_tag": {
            "type": "string",
            "title": "Explanation Tag"
          },
          "explanation": {
            "items": {
              "$ref": "#/components/schemas/ExplanationOut"
            },
            "type": "array",
            "title": "Explanation"
          },
          "why": {
            "$ref": "#/components/schemas/InsightWhyOut"
          },
          "small_step": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StepOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "learning_unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LearningUnitOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "action": {
            "$ref": "#/components/schemas/ActionOut"
          },
          "chat": {
            "$ref": "#/components/schemas/ChatOut"
          },
          "image": {
            "$ref": "#/components/schemas/InsightImageOut"
          },
          "publication": {
            "$ref": "#/components/schemas/PublicationOut"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "place_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[1-9][0-9]{0,18}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "disclosure": {
            "type": "string",
            "title": "Disclosure"
          }
        },
        "type": "object",
        "required": [
          "id",
          "scan_id",
          "origin",
          "engine",
          "label",
          "title",
          "glimpse",
          "anchor",
          "relation",
          "relation_label",
          "quran",
          "hadith",
          "hadith_status",
          "notice",
          "pair_complete",
          "explanation_tag",
          "explanation",
          "why",
          "small_step",
          "learning_unit",
          "action",
          "chat",
          "image",
          "publication",
          "completed_at",
          "place_id",
          "created_at",
          "disclosure"
        ],
        "title": "InsightDetailOut"
      },
...
      "InsightOrigin": {
        "type": "string",
        "enum": ["scan", "tutorial"],
        "title": "InsightOrigin"
      },
...
      "PublicInsightOut": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$",
            "title": "Id"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "glimpse": {
            "type": "string",
            "title": "Glimpse"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label",
            "description": "«مثال موثّق مُعدّ» when the insight is a prepared one"
          },
          "relation": {
            "$ref": "#/components/schemas/RelationType"
          },
          "relation_label": {
            "type": "string",
            "title": "Relation Label"
          },
          "quran": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InsightQuran"
              },
              {
                "type": "null"
              }
            ]
          },
          "hadith": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InsightHadith"
              },
              {
                "type": "null"
              }
            ]
          },
          "explanation_tag": {
            "type": "string",
            "title": "Explanation Tag"
          },
          "explanation": {
            "items": {
              "$ref": "#/components/schemas/ExplanationOut"
            },
            "type": "array",
            "title": "Explanation"
          },
          "why": {
            "$ref": "#/components/schemas/PublicWhyOut"
          },
          "small_step": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StepOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "author": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublicAuthorOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          },
          "disclosure": {
            "type": "string",
            "title": "Disclosure"
          }
        },
        "type": "object",
        "required": [
          "id",
          "path",
          "title",
          "glimpse",
          "label",
          "relation",
          "relation_label",
          "quran",
          "hadith",
          "explanation_tag",
          "explanation",
          "why",
          "small_step",
          "author",
          "published_at",
          "disclosure"
        ],
        "title": "PublicInsightOut",
        "description": "A published insight as anyone reads it.\n\nThe texts shown, read from the store by reference, and the platform's own words;\nnothing of the person: no profile, no chat, no learning history, no scan, no place\nand no photo (v2 §18)."
      },
...
      "PublicationOut": {
        "properties": {
          "published": {
            "type": "boolean",
            "title": "Published"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path",
            "description": "The public page's path on the web app, while published"
          }
        },
        "type": "object",
        "required": ["published", "published_at", "path"],
        "title": "PublicationOut",
        "description": "Whether the owner made the insight public (v2 §18), since when, and where anyone reads it."
      },
...
      "RelationType": {
        "type": "string",
        "enum": ["direct", "action_based", "close_conceptual", "opposite", "thematic_reminder"],
        "title": "RelationType",
        "description": "How an insight relates to the scene, strongest first (v2 §8)."
      },
...
    }
  },
...
}